requirements.txt
streamlit
pandas
matplotlib
seaborn
plotly.express
scikit-learn
streamlit_option_menu